split and show two stack bar separately
