+++ src/App.js
@@ -1,48 +1,53 @@
 import './App.css';
-import React, { useState } from 'react';
+import React, {useState} from 'react';
 import Select from 'react-select';
 import CalendarTable from './CalendarTable'
 import HeatMap from './HeatMap'
-import StackedChart from './StackedChart'
+import AirflowStackedChart from './AirflowStackedChart'
+import TemperatureStackedChart from "./TemperatureStackedChart";
 
 const Views = [
-  { label: "Pie Chart", value: <CalendarTable /> },
-  { label: "Stacked Bar", value: <StackedChart /> },
-  { label: "Heat Map", value: <HeatMap title={"Heat Map of ZoneTemperature"} labels={["Summer Occurrence", "Controlled Summer", "Controlled Winter", "Override Release"]} /> } /*temp place holder */
+    {label: "Pie Chart", value: <CalendarTable/>},
+    {label: "Stacked Bar(AirFlow)", value: <AirflowStackedChart/>},
+    {label: "Stacked Bar(Temperature)", value: <TemperatureStackedChart/>},
+    {label: "Heat Map",
+        value: <HeatMap title={"Heat Map of ZoneTemperature"}
+                        labels={["Summer Occurrence", "Controlled Summer", "Controlled Winter", "Override Release"]}/>
+    } /*temp place holder */
 ];
 
 
 function App(): React.Component {
 
-  const [selectedView, setView] = useState(<CalendarTable />);
+    const [selectedView, setView] = useState(<CalendarTable/>);
 
-  const handleChange = (event) => {
-      setView(event.value);
-  };
-  
-  return (
-    <div>
-      <div className="App">
-        <header className="App-header">
-        <p >SEB Building</p>
-        </header>
-      </div>
-      <div style={{width: '200px'}}>
-        <Select placeholder="Select View"
-          options={Views}
-	  defaultValue = {{ label: "Pie Chart", value: <CalendarTable /> }}
-          onChange={handleChange} />
-      </div>
-      <div>
-	      <p> {selectedView} </p>
-      </div>
+    const handleChange = (event) => {
+        setView(event.value);
+    };
+
+    return (
         <div>
-            <iframe
-                src="https://ec2-3-140-252-185.us-east-2.compute.amazonaws.com:3000/d/pSc0RlYMzd/pnnl-seb-influx?orgId=1"
-                width="100%" height="700px"></iframe>
+            <div className="App">
+                <header className="App-header">
+                    <p>SEB Building</p>
+                </header>
+            </div>
+            <div style={{width: '200px'}}>
+                <Select placeholder="Select View"
+                        options={Views}
+                        defaultValue={{label: "Pie Chart", value: <CalendarTable/>}}
+                        onChange={handleChange}/>
+            </div>
+            <div>
+                <p> {selectedView} </p>
+            </div>
+            <div>
+                <iframe
+                    src="https://ec2-3-140-252-185.us-east-2.compute.amazonaws.com:3000/d/pSc0RlYMzd/pnnl-seb-influx?orgId=1"
+                    width="100%" height="700px"></iframe>
+            </div>
         </div>
-    </div>
-  );
+    );
 }
 
 export default App;
